tests/wpt/web-platform-tests/html/semantics/scripting-1/the-script-element/execution-timing/023.html

<!DOCTYPE html>
<html><head>
        <title> scheduler: DOM added script, even later .src </title>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <script src="testlib/testlib.js"></script>
</head>
<body>

        <div id="log">FAILED (This TC requires JavaScript enabled)</div>
        <script>log('inline script #1');
                var script = testlib.addScript('', { }, document.getElementsByTagName('body')[0], false);
                log('end script #1');
        </script>

        <script type="text/javascript">
        log( 'inline script #2' );
        var t = async_test()

        function test() {
                assert_array_equals(eventOrder, ['inline script #1', 'end script #1', 'inline script #2', 'external script #1']);
                t.done();
}
        onload = t.step_func(function(){
                script.src='scripts/include-1.js';
                script.onload = t.step_func(test);
        })
        </script>

</body></html>
